@@ -2,7 +2,7 @@ import groovy.json.JsonSlurper
 
 // Top-level build file where you can add configuration options common to all sub-projects/modules.
 buildscript {
-    ext.kotlin_version = '1.4.10'
+    ext.kotlin_version = '1.6.10'
     repositories {
         mavenLocal()
         maven { url 'https://maven.aliyun.com/repository/central' }
@@ -16,8 +16,8 @@ buildscript {
     dependencies {
         classpath 'com.android.tools.build:gradle:4.1.0'
         classpath "org.jetbrains.kotlin:kotlin-gradle-plugin:$kotlin_version"
-        classpath 'com.jakewharton:butterknife-gradle-plugin:10.2.1'
-        classpath 'org.codehaus.groovy:groovy-json:3.0.6'
+        classpath 'com.jakewharton:butterknife-gradle-plugin:10.2.3'
+        classpath 'org.codehaus.groovy:groovy-json:3.0.8'
 
 
     }
