@@ -3,11 +3,15 @@ package org.autojs.autojs.ui.main;
 import android.animation.Animator;
 import android.animation.AnimatorListenerAdapter;
 import android.content.Context;
+
 import androidx.annotation.AttrRes;
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
+
 import com.google.android.material.floatingactionbutton.FloatingActionButton;
+
 import androidx.interpolator.view.animation.FastOutSlowInInterpolator;
+
 import android.util.AttributeSet;
 import android.view.LayoutInflater;
 import android.view.View;
@@ -30,11 +34,12 @@ public class FloatingActionMenu extends FrameLayout implements View.OnClickListe
     }
 
     private static final int[] ICONS = {
+            R.drawable.ic_floating_action_menu_home,
+            R.drawable.ic_floating_action_menu_open,
             R.drawable.ic_floating_action_menu_dir,
             R.drawable.ic_floating_action_menu_file,
-            R.drawable.ic_floating_action_menu_open,
             R.drawable.ic_project};
-    private static final int[] LABELS = {R.string.text_directory, R.string.text_file, R.string.text_import, R.string.text_project};
+    private static final int[] LABELS = {R.string.text_homepage, R.string.text_import, R.string.text_directory, R.string.text_file, R.string.text_project};
     private TextView[] mLabels;
     private FloatingActionButton[] mFabs;
     private View[] mFabContainers;
