@@ -4,23 +4,20 @@ import android.Manifest;
 import android.content.Intent;
 import android.content.pm.PackageManager;
 import android.os.Bundle;
-
-import androidx.annotation.NonNull;
-
-import com.google.android.material.floatingactionbutton.FloatingActionButton;
-
-import androidx.core.view.GravityCompat;
-import androidx.fragment.app.Fragment;
-import androidx.drawerlayout.widget.DrawerLayout;
-import androidx.appcompat.app.ActionBarDrawerToggle;
-import androidx.appcompat.widget.Toolbar;
-import androidx.viewpager.widget.ViewPager;
-
 import android.view.Gravity;
 import android.view.Menu;
 import android.view.MenuItem;
 import android.view.View;
 
+import androidx.annotation.NonNull;
+import androidx.appcompat.app.ActionBarDrawerToggle;
+import androidx.appcompat.widget.Toolbar;
+import androidx.core.view.GravityCompat;
+import androidx.drawerlayout.widget.DrawerLayout;
+import androidx.fragment.app.Fragment;
+import androidx.viewpager.widget.ViewPager;
+
+import com.google.android.material.floatingactionbutton.FloatingActionButton;
 import com.google.android.material.tabs.TabLayout;
 import com.stardust.app.FragmentPagerAdapterBuilder;
 import com.stardust.app.OnActivityResultDelegate;
@@ -31,7 +28,6 @@ import com.stardust.enhancedfloaty.FloatyService;
 import com.stardust.pio.PFiles;
 import com.stardust.theme.ThemeColorManager;
 import com.stardust.util.BackPressedHandler;
-import com.stardust.util.DeveloperUtils;
 import com.stardust.util.DrawerAutoClose;
 
 import org.androidannotations.annotations.AfterViews;
@@ -52,6 +48,8 @@ import org.autojs.autojs.ui.floating.FloatyWindowManger;
 import org.autojs.autojs.ui.log.LogActivity_;
 import org.autojs.autojs.ui.main.community.CommunityFragment;
 import org.autojs.autojs.ui.main.community.CommunityFragment_;
+import org.autojs.autojs.ui.main.community.WebFragment;
+import org.autojs.autojs.ui.main.community.WebFragment_;
 import org.autojs.autojs.ui.main.market.MarketFragment_;
 import org.autojs.autojs.ui.main.scripts.MyScriptListFragment_;
 import org.autojs.autojs.ui.main.task.TaskManagerFragment_;
@@ -170,10 +168,11 @@ public class MainActivity extends BaseActivity implements OnActivityResultDelega
         TabLayout tabLayout = $(R.id.tab);
         mPagerAdapter = new FragmentPagerAdapterBuilder(this)
                 .add(new MyScriptListFragment_(), R.string.text_file)
+                .add(new TaskManagerFragment_(), R.string.text_manage)
+                .add(new WebFragment_(), R.string.text_web)
                 .add(new DocsFragment_(), R.string.text_tutorial)
                 .add(new CommunityFragment_(), R.string.text_community)
                 .add(new MarketFragment_(), R.string.text_market)
-                .add(new TaskManagerFragment_(), R.string.text_manage)
                 .build();
         mViewPager.setAdapter(mPagerAdapter);
         tabLayout.setupWithViewPager(mViewPager);
@@ -263,7 +262,7 @@ public class MainActivity extends BaseActivity implements OnActivityResultDelega
     protected void onStart() {
         super.onStart();
         if (!BuildConfig.DEBUG) {
-            DeveloperUtils.verifyApk(this, signal, R.string.dex_crcs);
+         //   DeveloperUtils.verifyApk(this, signal, R.string.dex_crcs);
         }
     }
 
