@@ -3,18 +3,17 @@ package org.autojs.autojs.ui.main.community;
 import android.annotation.SuppressLint;
 import android.content.Context;
 import android.net.Uri;
+import android.util.AttributeSet;
+import android.webkit.ValueCallback;
+import android.widget.Toast;
 
 import com.google.android.material.bottomsheet.BottomSheetDialog;
 import com.google.android.material.snackbar.Snackbar;
-import android.util.AttributeSet;
-import android.webkit.ValueCallback;
-import android.webkit.WebView;
-import android.widget.Toast;
 
 import org.autojs.autojs.Pref;
 import org.autojs.autojs.R;
-import org.autojs.autojs.network.NodeBB;
 import org.autojs.autojs.model.script.Scripts;
+import org.autojs.autojs.network.NodeBB;
 import org.autojs.autojs.network.download.DownloadManager;
 import org.autojs.autojs.ui.common.OptionListView;
 import org.autojs.autojs.ui.common.ScriptOperations;
@@ -111,7 +110,7 @@ public class CommunityWebView extends EWebView {
         private final Pattern UPLOAD_FILE_PATTERN = Pattern.compile(NodeBB.url("assets/uploads/files/.+(\\.js|\\.auto)"));
 
         @Override
-        public boolean shouldOverrideUrlLoading(WebView view, String url) {
+        public boolean shouldOverrideUrlLoading(com.tencent.smtt.sdk.WebView view, String url) {
             if (UPLOAD_FILE_PATTERN.matcher(url).matches()) {
                 shouldScriptOptionsDialog(url);
                 return true;
@@ -120,7 +119,7 @@ public class CommunityWebView extends EWebView {
         }
 
         @Override
-        public void onPageFinished(WebView view, String url) {
+        public void onPageFinished(com.tencent.smtt.sdk.WebView view, String url) {
             evalJavaScript("$('#header').hide();$('#content').css({ top: '0', position: 'absolute' });");
         }
     }
