@@ -53,7 +53,7 @@ runtime.init();
 
      //初始化不依赖环境的模块
      global.JSON = require('__json2__.js');
-     global.util = require('__util__.js');
+     global.util = global.$util = require('__util__.js');
      global.device = runtime.device;
      global.Promise = require('promise.js');
  
@@ -64,13 +64,17 @@ runtime.init();
     require("__globals__")(runtime, global);
     //初始化一般模块
     (function (scope) {
-        var modules = ['app', 'automator', 'console', 'dialogs', 'io', 'selector', 'shell', 'web', 'ui',
+        var modules = ['app', 'automator', 'console', 'dialogs', 'files', 'io', 'selector', 'shell', 'web', 'ui',
             "images", "threads", "events", "engines", "RootAutomator", "http", "storages", "floaty",
-            "sensors", "media", "plugins", "continuation"];
+            "sensors", "media", "plugins", "continuation", "$zip", "$base64", "$crypto"];
         var len = modules.length;
         for (var i = 0; i < len; i++) {
             var m = modules[i];
-            scope[m] = require('__' + m + '__')(scope.runtime, scope);
+            let module = require('__' + m + '__')(scope.runtime, scope);
+            scope[m] = module;
+            if(!m.startsWith('$')) {
+                scope['$' + m] = module;
+            }
         }
     })(global);
 
@@ -79,6 +83,8 @@ runtime.init();
     importClass(android.graphics.Paint);
     Canvas = com.stardust.autojs.core.graphics.ScriptCanvas;
     Image = com.stardust.autojs.core.image.ImageWrapper;
+    OkHttpClient = Packages["okhttp3"].OkHttpClient;
+    Intent = android.content.Intent;
 
     //重定向require以便支持相对路径和npm模块
     Module = require("jvm-npm.js");
