@@ -53,7 +53,7 @@ public class OCRPredictorNative {
 
     }
 
-    public void destory(){
+    public void destroy(){
         if (nativePointer != 0) {
             release(nativePointer);
             nativePointer = 0;
@@ -98,5 +98,9 @@ public class OCRPredictorNative {
         return model;
     }
 
-
+    @Override
+    protected void finalize() throws Throwable {
+        super.finalize();
+        destroy();
+    }
 }
