@@ -167,7 +167,7 @@ public class ConsoleImpl extends AbstractConsole {
     }
 
     @Override
-    public void setBackgroud(@Nullable String color) {
+    public void setBackground(@Nullable String color) {
         if(mConsoleView.get()==null){
             Log.e(ConsoleImpl.class.getName(), "设置不生效，console没创建创建 ");
             return;
