@@ -57,6 +57,122 @@ class GlobalActionAutomator(private val mHandler: Handler?, private val serviceP
         return performGlobalAction(AccessibilityService.GLOBAL_ACTION_RECENTS)
     }
 
+    /**
+     * Action to take a screenshot
+     */
+    @RequiresApi(Build.VERSION_CODES.P)
+    fun takeScreenshot(): Boolean {
+        return performGlobalAction(AccessibilityService.GLOBAL_ACTION_TAKE_SCREENSHOT)
+    }
+
+    /**
+     * Action to lock the screen
+     */
+    @RequiresApi(Build.VERSION_CODES.P)
+    fun lockScreen(): Boolean {
+        return performGlobalAction(AccessibilityService.GLOBAL_ACTION_LOCK_SCREEN)
+    }
+
+    /**
+     * Action to dismiss the notification shade
+     */
+    @RequiresApi(Build.VERSION_CODES.S)
+    fun dismissNotificationShade(): Boolean {
+        return performGlobalAction(AccessibilityService.GLOBAL_ACTION_DISMISS_NOTIFICATION_SHADE)
+    }
+
+    /**
+     * Action to send the KEYCODE_HEADSETHOOK KeyEvent, which is used to answer/hang up
+     * calls and play/stop media
+     */
+    @RequiresApi(Build.VERSION_CODES.S)
+    fun keyCodeHeadsetHook(): Boolean {
+        return performGlobalAction(AccessibilityService.GLOBAL_ACTION_KEYCODE_HEADSETHOOK)
+    }
+
+    /**
+     * Action to trigger the Accessibility Shortcut. This shortcut has a hardware trigger
+     * and can be activated by holding down the two volume keys.
+     */
+    @RequiresApi(Build.VERSION_CODES.S)
+    fun accessibilityShortcut(): Boolean {
+        return performGlobalAction(AccessibilityService.GLOBAL_ACTION_ACCESSIBILITY_SHORTCUT)
+    }
+
+    /**
+     * Action to bring up the Accessibility Button’s chooser menu
+     */
+    @RequiresApi(Build.VERSION_CODES.S)
+    fun accessibilityButtonChooser(): Boolean {
+        return performGlobalAction(AccessibilityService.GLOBAL_ACTION_ACCESSIBILITY_BUTTON_CHOOSER)
+    }
+
+    /**
+     * Action to trigger the Accessibility Button
+     */
+    @RequiresApi(Build.VERSION_CODES.S)
+    fun accessibilityButton(): Boolean {
+        return performGlobalAction(AccessibilityService.GLOBAL_ACTION_ACCESSIBILITY_BUTTON)
+    }
+
+    /**
+     * Action to show Launcher’s all apps.
+     */
+    @RequiresApi(Build.VERSION_CODES.S)
+    fun accessibilityAllApps(): Boolean {
+        return performGlobalAction(AccessibilityService.GLOBAL_ACTION_ACCESSIBILITY_ALL_APPS)
+    }
+
+    /**
+     * Action to trigger dpad up keyevent.
+     */
+//    @RequiresApi(Build.VERSION_CODES.Tiramisu)
+    fun dpadUp(): Boolean {
+//        return performGlobalAction(AccessibilityService.GLOBAL_ACTION_DPAD_UP)
+        // TODO: 待适配Api Tiramisu
+        return false
+    }
+
+    /**
+     * Action to trigger dpad down keyevent.
+     */
+//    @RequiresApi(Build.VERSION_CODES.Tiramisu)
+    fun dpadDown(): Boolean {
+//        return performGlobalAction(AccessibilityService.GLOBAL_ACTION_DPAD_DOWN)
+        // TODO: 待适配Api Tiramisu
+        return false
+    }
+
+    /**
+     * Action to trigger dpad right keyevent.
+     */
+//    @RequiresApi(Build.VERSION_CODES.Tiramisu)
+    fun dpadRight(): Boolean {
+//        return performGlobalAction(AccessibilityService.GLOBAL_ACTION_DPAD_RIGHT)
+        // TODO: 待适配Api Tiramisu
+        return false
+    }
+
+    /**
+     * Action to trigger dpad left keyevent.
+     */
+//    @RequiresApi(Build.VERSION_CODES.Tiramisu)
+    fun dpadLeft(): Boolean {
+//        return performGlobalAction(AccessibilityService.GLOBAL_ACTION_DPAD_LEFT)
+        // TODO: 待适配Api Tiramisu
+        return false
+    }
+
+    /**
+     * Action to trigger dpad center keyevent.
+     */
+//    @RequiresApi(Build.VERSION_CODES.Tiramisu)
+    fun dpadCenter(): Boolean {
+//        return performGlobalAction(AccessibilityService.GLOBAL_ACTION_DPAD_CENTER)
+        // TODO: 待适配Api Tiramisu
+        return false
+    }
+
     @RequiresApi(api = Build.VERSION_CODES.N)
     fun splitScreen(): Boolean {
         return performGlobalAction(AccessibilityService.GLOBAL_ACTION_TOGGLE_SPLIT_SCREEN)
@@ -117,7 +233,7 @@ class GlobalActionAutomator(private val mHandler: Handler?, private val serviceP
     private fun gesturesWithoutHandler(description: GestureDescription): Boolean {
         prepareLooperIfNeeded()
         val result = VolatileBox(false)
-        val handler = Handler(Looper.myLooper())
+        val handler = Looper.myLooper()?.let { Handler(it) }
         service.dispatchGesture(description, object : AccessibilityService.GestureResultCallback() {
             override fun onCompleted(gestureDescription: GestureDescription) {
                 result.set(true)
