@@ -83,8 +83,8 @@ open class AccessibilityNodeInfoAllocator {
             try {
                 key.recycle()
                 notRecycledCount++
-                if (DEBUG)
-                    Log.w(TAG, value)
+//                 if (DEBUG)
+//                     Log.w(TAG, value)
             } catch (ignored: IllegalStateException) {
             }
 
