@@ -16,11 +16,9 @@ import androidx.annotation.RequiresApi
 import androidx.appcompat.app.AppCompatActivity
 import com.afollestad.materialdialogs.MaterialDialog
 import com.linsh.utilseverywhere.IntentUtils
-import com.stardust.app.GlobalAppContext
 import com.stardust.auojs.inrt.autojs.AutoJs
 import com.stardust.auojs.inrt.launch.GlobalProjectLauncher
 import com.stardust.auojs.inrt.util.UpdateUtil
-import com.stardust.autojs.project.ProjectConfig
 import com.stardust.util.IntentUtil
 import ezy.assist.compat.SettingsCompat
 import java.util.*
@@ -38,40 +36,28 @@ class SplashActivity : AppCompatActivity() {
         setContentView(R.layout.activity_splash)
         val slug = findViewById<TextView>(R.id.slug)
         slug.typeface = Typeface.createFromAsset(assets, "roboto_medium.ttf")
-        if(Pref.getHost("d")=="d"){ //非第一次运行
-            Pref.setHost("112.74.161.35")
-            val mProjectConfig: ProjectConfig = ProjectConfig.fromAssets(this, ProjectConfig.configFileOfDir("project"))
-            Pref.setHideLogs(mProjectConfig.getLaunchConfig().shouldHideLogs())
-            Pref.setStableMode(mProjectConfig.getLaunchConfig().isStableMode())
-            Pref.setStopAllScriptsWhenVolumeUp(mProjectConfig.getLaunchConfig().isVolumeUpcontrol())
-            Pref.setDisplaySplash(mProjectConfig.getLaunchConfig().isDisplaySplash())
-        }
-        if (!BuildConfig.isMarket) {
-            if(Pref.istDisplaySplash()){
-                main()
-            }else{
+        if (!Pref.isFirstUsing()) {//不是第一次
+            main()
+        } else {
+            if(Pref.getHost("d")=="d"){
+                Pref.setHost("112.74.161.35")
+            }
+            if (!BuildConfig.isMarket) {
                 Handler().postDelayed({ this@SplashActivity.main() }, INIT_TIMEOUT)
             }
-        }else{
-           main()
         }
-
     }
 
 
     private fun main() {
-        if(checkPermission(Manifest.permission.WRITE_EXTERNAL_STORAGE,
-                Manifest.permission.READ_PHONE_STATE)){
-            runScript();
-        }else{
-            GlobalAppContext.toast("请开启权限后，再运行!")
-        }
+        checkPermission(Manifest.permission.WRITE_EXTERNAL_STORAGE,
+                Manifest.permission.READ_PHONE_STATE)
     }
 
     private fun manageDrawOverlays() {
         var dialog = MaterialDialog.Builder(this).title("提示").content("请打开所有的权限，\r\n 省电策略选【不限制】")//内容
                 .positiveText("确定") //肯定按键
-                .onPositive { _, _ ->
+                .onPositive { dialog, which ->
                     SettingsCompat.manageDrawOverlays(this);
                 }.canceledOnTouchOutside(false)
                 .build();
@@ -81,7 +67,7 @@ class SplashActivity : AppCompatActivity() {
     private fun manageWriteSettings() {
         var dialog = MaterialDialog.Builder(this).title("继续进入权限设置").content("请打开所有权限!\r\n 请打开所有权限 \r\n 请打开所有权限")//内容
                 .positiveText("确定") //肯定按键
-                .onPositive { _, _ ->
+                .onPositive { dialog, which ->
                     IntentUtil.goToAppDetailSettings(this);
                 }.canceledOnTouchOutside(false)
                 .build();
@@ -108,6 +94,8 @@ class SplashActivity : AppCompatActivity() {
         }
         Thread {
             try {
+
+                Thread.sleep(2000);
                 GlobalProjectLauncher.launch(this)
                 this.finish();
             } catch (e: Exception) {
@@ -115,7 +103,7 @@ class SplashActivity : AppCompatActivity() {
                 runOnUiThread {
                     Toast.makeText(this@SplashActivity, e.message, Toast.LENGTH_LONG).show()
                     startActivity(Intent(this@SplashActivity, LogActivity::class.java))
-                    AutoJs.instance.globalConsole.printAllStackTrace(e)
+                    AutoJs.instance!!.globalConsole.printAllStackTrace(e)
                 }
             }
         }.start()
@@ -123,19 +111,19 @@ class SplashActivity : AppCompatActivity() {
 
     override fun onRequestPermissionsResult(requestCode: Int, @NonNull permissions: Array<String>, @NonNull grantResults: IntArray) {
         Log.d(TAG, "onRequestPermissionsResult: " + requestCode);
+        runScript()
     }
 
-    private fun checkPermission(vararg permissions: String): Boolean {
+    private fun checkPermission(vararg permissions: String) {
         if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.M) {
             val requestPermissions = getRequestPermissions(permissions)
             if (requestPermissions.isNotEmpty()) {
                 requestPermissions(requestPermissions, PERMISSION_REQUEST_CODE)
-                return false;
             } else {
-              return true;
+                runScript()
             }
         } else {
-            return true;
+            runScript()
         }
     }
 
@@ -159,7 +147,8 @@ class SplashActivity : AppCompatActivity() {
 
     override fun onResume() {
         if (BuildConfig.isMarket) {
-            if (Pref.isFirstUsing()) { //已经不是第一次了
+            Log.d(TAG, "onResume: ")
+            if (!Pref.isFirstUsing()) { //已经不是第一次了
                 if (step == 1) {
                     manageDrawOverlays();
                 }
@@ -170,13 +159,14 @@ class SplashActivity : AppCompatActivity() {
                     AccessibilitySetting();
                 }
                 if (step == 4) {
-                    Pref.setNotFirstUsingEnd()
                     main();
                 }
                 step++;
             }
+
         }
         super.onResume()
+
     }
 
 }
